# apps/example/public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Next REST Framework",
    "description": "This is an autogenerated documentation by Next REST Framework.",
    "version": "v4.3.0"
  },
  "paths": {
    "/api/v1/rpc/createTodo": {
      "post": {
        "operationId": "createTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTodoRequestBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CreateTodoResponseBody",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodoResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["RPC"]
      }
    },
    "/api/v1/rpc/deleteTodo": {
      "post": {
        "operationId": "deleteTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeleteTodoRequestBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DeleteTodoResponseBody2",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTodoResponseBody2"
                }
              }
            }
          },
          "400": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["RPC"]
      }
    },
    "/api/v1/rpc/getTodoById": {
      "post": {
        "operationId": "getTodoById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTodoByIdRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "GetTodoByIdResponseBody2",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoByIdResponseBody2"
                }
              }
            }
          },
          "400": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["RPC"]
      }
    },
    "/api/v1/rpc/getTodos": {
      "post": {
        "operationId": "getTodos",
        "responses": {
          "200": {
            "description": "GetTodosResponseBody",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodosResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["RPC"]
      }
    },
    "/api/v1/todos": {
      "get": {
        "operationId": "getTodos",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodos200ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["example-api", "todos", "pages-router"]
      },
      "post": {
        "operationId": "createTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTodoRequestBody" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for status 201",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodo201ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodo401ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["example-api", "todos", "pages-router"]
      }
    },
    "/api/v1/todos/{id}": {
      "get": {
        "operationId": "getTodoById",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoById200ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoById404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["example-api", "todos", "pages-router"]
      },
      "delete": {
        "operationId": "deleteTodo",
        "responses": {
          "204": {
            "description": "Response for status 204",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTodo204ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTodo404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["example-api", "todos", "pages-router"]
      }
    },
    "/api/v2/rpc/createTodo": {
      "post": {
        "operationId": "createTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTodoRequestBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CreateTodoResponseBody",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodoResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["RPC"]
      }
    },
    "/api/v2/rpc/deleteTodo": {
      "post": {
        "operationId": "deleteTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeleteTodoRequestBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DeleteTodoResponseBody2",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTodoResponseBody2"
                }
              }
            }
          },
          "400": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["RPC"]
      }
    },
    "/api/v2/rpc/getTodoById": {
      "post": {
        "operationId": "getTodoById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTodoByIdRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "GetTodoByIdResponseBody2",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoByIdResponseBody2"
                }
              }
            }
          },
          "400": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["RPC"]
      }
    },
    "/api/v2/rpc/getTodos": {
      "post": {
        "operationId": "getTodos",
        "responses": {
          "200": {
            "description": "GetTodosResponseBody",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodosResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["RPC"]
      }
    },
    "/api/v2/todos": {
      "get": {
        "operationId": "getTodos",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodos200ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["example-api", "todos", "app-router"]
      },
      "post": {
        "operationId": "createTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTodoRequestBody" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for status 201",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodo201ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodo401ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "tags": ["example-api", "todos", "app-router"]
      }
    },
    "/api/v2/todos/{id}": {
      "get": {
        "operationId": "getTodoById",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoById200ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoById404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["example-api", "todos", "app-router"]
      },
      "delete": {
        "operationId": "deleteTodo",
        "responses": {
          "204": {
            "description": "Response for status 204",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTodo204ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTodo404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["example-api", "todos", "app-router"]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateTodo201ResponseBody": { "type": "string" },
      "CreateTodo401ResponseBody": { "type": "string" },
      "CreateTodoRequestBody": {
        "type": "object",
        "properties": { "name": { "type": "string" } },
        "required": ["name"],
        "additionalProperties": false
      },
      "CreateTodoResponseBody": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "completed": { "type": "boolean" }
        },
        "required": ["id", "name", "completed"],
        "additionalProperties": false
      },
      "DeleteTodo204ResponseBody": { "type": "string" },
      "DeleteTodo404ResponseBody": { "type": "string" },
      "DeleteTodoRequestBody": { "type": "string" },
      "DeleteTodoResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "DeleteTodoResponseBody2": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"],
        "additionalProperties": false
      },
      "GetTodoById200ResponseBody": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "completed": { "type": "boolean" }
        },
        "required": ["id", "name", "completed"],
        "additionalProperties": false
      },
      "GetTodoById404ResponseBody": { "type": "string" },
      "GetTodoByIdRequestBody": { "type": "string" },
      "GetTodoByIdResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetTodoByIdResponseBody2": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "completed": { "type": "boolean" }
        },
        "required": ["id", "name", "completed"],
        "additionalProperties": false
      },
      "GetTodos200ResponseBody": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": { "type": "number" },
            "name": { "type": "string" },
            "completed": { "type": "boolean" }
          },
          "required": ["id", "name", "completed"],
          "additionalProperties": false
        }
      },
      "GetTodosResponseBody": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": { "type": "number" },
            "name": { "type": "string" },
            "completed": { "type": "boolean" }
          },
          "required": ["id", "name", "completed"],
          "additionalProperties": false
        }
      },
      "UnexpectedError": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "additionalProperties": false
      }
    }
  }
}